+++ themes/10up-block-theme/composer.json
@@ -20,7 +20,7 @@
   },
   "autoload": {
     "psr-4": {
-      "TenupBlockTheme\\": "includes/classes/"
+      "TenupBlockTheme\\": "src/"
     }
   },
   "config": {
