+++ themes/10up-block-theme/src/Blocks.php
@@ -1,0 +1,124 @@
+<?php
+/**
+ * Gutenberg Blocks setup
+ *
+ * @package TenupBlockTheme
+ */
+
+namespace TenupBlockTheme;
+
+use TenupFramework\Module;
+use TenupFramework\ModuleInterface;
+use TenupBlockTheme\Traits\GetAssetInfo;
+
+/**
+ * Blocks module.
+ *
+ * @package TenupBlockTheme
+ */
+class Blocks implements ModuleInterface {
+
+	use Module;
+	use GetAssetInfo;
+
+	/**
+	 * Can this module be registered?
+	 *
+	 * @return bool
+	 */
+	public function can_register() {
+		return true;
+	}
+
+	/**
+	 * Register any hooks and filters.
+	 *
+	 * @return void
+	 */
+	public function register() {
+		add_action( 'init', [ $this, 'register_theme_blocks' ], 10, 0 );
+		add_action( 'init', [ $this, 'enqueue_theme_block_styles' ], 10, 0 );
+	}
+
+
+	/**
+	 * Automatically registers all blocks that are located within the includes/blocks directory
+	 *
+	 * @return void
+	 */
+	public function register_theme_blocks() {
+		// Register all the blocks in the theme.
+		if ( file_exists( TENUP_BLOCK_THEME_BLOCK_DIST_DIR ) ) {
+			$block_json_files = glob( TENUP_BLOCK_THEME_BLOCK_DIST_DIR . '*/block.json' );
+			$block_names      = [];
+
+			if ( empty( $block_json_files ) ) {
+				return;
+			}
+
+			foreach ( $block_json_files as $filename ) {
+				$block_folder = dirname( $filename );
+				$block        = register_block_type_from_metadata( $block_folder );
+
+				if ( ! $block ) {
+					continue;
+				}
+
+				$block_names[] = $block->name;
+			}
+
+			add_filter(
+				'allowed_block_types_all',
+				function ( array|bool $allowed_blocks ) use ( $block_names ): array|bool {
+					if ( ! is_array( $allowed_blocks ) ) {
+						return $allowed_blocks;
+					}
+					return array_merge( $allowed_blocks, $block_names );
+				}
+			);
+		}
+	}
+
+	/**
+	 * Enqueue block specific styles.
+	 *
+	 * @return void
+	 */
+	public function enqueue_theme_block_styles() {
+		$stylesheets = glob( TENUP_BLOCK_THEME_DIST_PATH . '/blocks/autoenqueue/**/*.css' );
+
+		if ( empty( $stylesheets ) ) {
+			return;
+		}
+
+		foreach ( $stylesheets as $stylesheet_path ) {
+			$block_type = str_replace( TENUP_BLOCK_THEME_DIST_PATH . '/blocks/autoenqueue/', '', $stylesheet_path );
+			$block_type = str_replace( '.css', '', $block_type );
+
+			wp_register_style(
+				"tenup-theme-{$block_type}",
+				TENUP_BLOCK_THEME_DIST_URL . 'blocks/autoenqueue/' . $block_type . '.css',
+				$this->get_asset_info( 'blocks/autoenqueue/' . $block_type, 'dependencies' ),
+				$this->get_asset_info( 'blocks/autoenqueue/' . $block_type, 'version' ),
+			);
+
+			wp_enqueue_block_style(
+				$block_type,
+				[
+					'handle' => "tenup-theme-{$block_type}",
+					'path'   => $stylesheet_path,
+				]
+			);
+
+			if ( file_exists( TENUP_BLOCK_THEME_DIST_PATH . 'blocks/autoenqueue/' . $block_type . '.js' ) ) {
+				wp_enqueue_script(
+					$block_type,
+					TENUP_BLOCK_THEME_DIST_URL . 'blocks/autoenqueue/' . $block_type . '.js',
+					$this->get_asset_info( 'blocks/autoenqueue/' . $block_type, 'dependencies' ),
+					$this->get_asset_info( 'blocks/autoenqueue/' . $block_type, 'version' ),
+					true
+				);
+			}
+		}
+	}
+}
